Convert worker pool modules to ESM
Closes #5969.
