+++ lib/mocha.js
@@ -1118,7 +1118,7 @@ Mocha.prototype.parallelMode = function parallelMode(enable = true) {
 
   // swap Runner class
   this._runnerClass = parallel
-    ? require("./nodejs/parallel-buffered-runner")
+    ? require("./nodejs/parallel-buffered-runner.mjs").ParallelBufferedRunner
     : exports.Runner;
 
   // lazyLoadFiles may have been set `true` otherwise (for ESM loading),
