+++ lib/nodejs/parallel-buffered-runner.mjs
@@ -4,7 +4,12 @@
  * @private
  */
 
-"use strict";
+import { fileURLToPath } from "node:url";
+import debugFactory from "debug";
+import Runner from "../runner.js";
+import { createFatalError } from "../errors.mjs";
+import utils from "../utils.js";
+import { BufferedWorkerPool } from "./buffered-worker-pool.mjs";
 
 /**
  * @typedef {import('../types.d.ts').FileRunner} FileRunner
@@ -13,17 +18,15 @@
  * @typedef {import('../types.d.ts').SigIntListener} SigIntListener
  */
 
-const Runner = require("../runner");
 const { EVENT_RUN_BEGIN, EVENT_RUN_END } = Runner.constants;
-const debug = require("debug")("mocha:parallel:parallel-buffered-runner");
-const { BufferedWorkerPool } = require("./buffered-worker-pool");
+const debug = debugFactory("mocha:parallel:parallel-buffered-runner");
 const { setInterval, clearInterval } = global;
-const { createMap, constants } = require("../utils");
+const { createMap, constants } = utils;
 const { MOCHA_ID_PROP_NAME } = constants;
-const { createFatalError } = require("../errors.mjs");
 
-const DEFAULT_WORKER_REPORTER =
-  require.resolve("./reporters/parallel-buffered");
+const DEFAULT_WORKER_REPORTER = fileURLToPath(
+  new URL("./reporters/parallel-buffered.mjs", import.meta.url),
+);
 
 /**
  * List of options to _not_ serialize for transmission to workers
@@ -81,7 +84,7 @@ const states = createMap({
  * {@link Runnable}s by itself!
  * @public
  */
-class ParallelBufferedRunner extends Runner {
+export class ParallelBufferedRunner extends Runner {
   constructor(...args) {
     super(...args);
 
@@ -418,5 +421,3 @@ class ParallelBufferedRunner extends Runner {
     return this;
   }
 }
-
-module.exports = ParallelBufferedRunner;
