+++ package.json
@@ -174,13 +174,14 @@
     "fs": false,
     "path": false,
     "supports-color": false,
-    "./lib/nodejs/buffered-worker-pool.js": false,
+    "./lib/nodejs/buffered-worker-pool.mjs": false,
     "./lib/nodejs/esm-utils.js": false,
     "./lib/nodejs/file-unloader.js": false,
-    "./lib/nodejs/parallel-buffered-runner.js": false,
+    "./lib/nodejs/parallel-buffered-runner.mjs": false,
     "./lib/nodejs/serializer.js": false,
-    "./lib/nodejs/worker.js": false,
-    "./lib/nodejs/reporters/parallel-buffered.js": false,
+    "./lib/nodejs/worker-core.mjs": false,
+    "./lib/nodejs/worker.mjs": false,
+    "./lib/nodejs/reporters/parallel-buffered.mjs": false,
     "./lib/cli/index.js": false
   },
   "overrides": {
