+++ test/node-unit/worker.spec.js
@@ -1,11 +1,9 @@
 "use strict";
 
 const serializeJavascript = require("serialize-javascript");
-const rewiremock = require("rewiremock/node");
 const { SerializableWorkerResult } = require("../../lib/nodejs/serializer");
 const sinon = require("sinon");
-
-const WORKER_PATH = require.resolve("../../lib/nodejs/worker.js");
+const { startWorker } = require("../../lib/nodejs/worker-core.mjs");
 
 describe("worker", function () {
   let worker;
@@ -24,7 +22,7 @@ describe("worker", function () {
   describe("when run as main process", function () {
     it("should throw", function () {
       expect(() => {
-        rewiremock.proxy(WORKER_PATH, {
+        startWorker({
           workerpool: {
             isMainThread: true,
             worker: stubs.workerpool.worker,
@@ -58,16 +56,12 @@ describe("worker", function () {
         validateLegacyPlugin: sinon.stub(),
       };
 
-      stubs.plugin = {
-        aggregateRootHooks: sinon.stub().resolves(),
-      };
-
-      worker = rewiremock.proxy(WORKER_PATH, {
+      worker = startWorker({
         workerpool: stubs.workerpool,
-        "../../lib/mocha": stubs.Mocha,
-        "../../lib/nodejs/serializer": stubs.serializer,
-        "../../lib/cli/run-helpers": stubs.runHelpers,
-        "../../lib/plugin-loader": stubs.plugin,
+        Mocha: stubs.Mocha,
+        handleRequires: stubs.runHelpers.handleRequires,
+        validateLegacyPlugin: stubs.runHelpers.validateLegacyPlugin,
+        serialize: stubs.serializer.serialize,
       });
     });
